pyproject.toml
[tool.poetry]
name = "gembrain"
version = "0.1.0"
description = "A Gemini-powered agentic second brain desktop application"
authors = ["GemBrain Contributors"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
PyQt6 = "^6.6.0"
SQLAlchemy = "^2.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
google-generativeai = "^0.3.0"
APScheduler = "^3.10.0"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-qt = "^4.2.0"
black = "^23.12.0"
ruff = "^0.1.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
gembrain = "gembrain.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
